ultimo cambio compras

+++ ms_compras/src/main/java/com/example/ms_compras/repository/ComprasRepository.java
@@ -3,5 +3,5 @@ package com.example.ms_compras.repository;
 import com.example.ms_compras.entity.Compras;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComprasRepository extends JpaRepository<Compras, Long> {
+public interface ComprasRepository extends JpaRepository<Compras, Integer> {
 }
